requirements.txt
anyio==3.6.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
fastapi==0.85.1
h11==0.14.0
idna==3.4
psutil==5.9.3
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
requests==2.28.1
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
